2_Cleaner/requirements.txt
numpy==2.3.5
pandas==2.3.3
azure-storage-blob==12.27.1